Assignment 3/ENSF607_Assignment3_Exercise2/app/requirements.txt
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
